Move testable code into separate file
